+++ idea/frontend/src/app/providers/account/consts.ts
@@ -1,0 +1,13 @@
+const DEFAULT_INJECT_TIMEOUT_MS = 200 as const;
+
+const WALLET_STATUS = {
+  INJECTED: 'injected',
+  CONNECTED: 'connected',
+} as const;
+
+const LOCAL_STORAGE_KEY = {
+  WALLET_IDS: 'walletIds',
+  ACCOUNT_ADDRESS: 'accountAddress',
+} as const;
+
+export { DEFAULT_INJECT_TIMEOUT_MS, WALLET_STATUS, LOCAL_STORAGE_KEY };
